+++ packages/twenty-server/src/modules/messaging/services/gmail-partial-message-list-fetch/gmail-get-history.service.ts
@@ -1,0 +1,104 @@
+import { Injectable } from '@nestjs/common';
+
+import { gmail_v1 } from 'googleapis';
+
+import { GMAIL_USERS_HISTORY_MAX_RESULT } from 'src/modules/messaging/constants/gmail-users-history-max-result.constant';
+import { GmailError } from 'src/modules/messaging/types/gmail-error';
+
+@Injectable()
+export class GmailGetHistoryService {
+  constructor() {}
+
+  public async getHistory(
+    gmailClient: gmail_v1.Gmail,
+    lastSyncHistoryId: string,
+  ): Promise<{
+    history: gmail_v1.Schema$History[];
+    historyId?: string | null;
+    error?: GmailError;
+  }> {
+    const fullHistory: gmail_v1.Schema$History[] = [];
+    let pageToken: string | undefined;
+    let hasMoreMessages = true;
+    let nextHistoryId: string | undefined;
+
+    while (hasMoreMessages) {
+      try {
+        const response = await gmailClient.users.history.list({
+          userId: 'me',
+          maxResults: GMAIL_USERS_HISTORY_MAX_RESULT,
+          pageToken,
+          startHistoryId: lastSyncHistoryId,
+          historyTypes: ['messageAdded', 'messageDeleted'],
+        });
+
+        nextHistoryId = response?.data?.historyId ?? undefined;
+
+        if (response?.data?.history) {
+          fullHistory.push(...response.data.history);
+        }
+
+        pageToken = response?.data?.nextPageToken ?? undefined;
+        hasMoreMessages = !!pageToken;
+      } catch (error) {
+        const errorData = error?.response?.data?.error;
+
+        if (errorData) {
+          return {
+            history: [],
+            error: errorData,
+            historyId: lastSyncHistoryId,
+          };
+        }
+
+        throw error;
+      }
+    }
+
+    return { history: fullHistory, historyId: nextHistoryId };
+  }
+
+  public async getMessageIdsFromHistory(
+    history: gmail_v1.Schema$History[],
+  ): Promise<{
+    messagesAdded: string[];
+    messagesDeleted: string[];
+  }> {
+    const { messagesAdded, messagesDeleted } = history.reduce(
+      (
+        acc: {
+          messagesAdded: string[];
+          messagesDeleted: string[];
+        },
+        history,
+      ) => {
+        const messagesAdded = history.messagesAdded?.map(
+          (messageAdded) => messageAdded.message?.id || '',
+        );
+
+        const messagesDeleted = history.messagesDeleted?.map(
+          (messageDeleted) => messageDeleted.message?.id || '',
+        );
+
+        if (messagesAdded) acc.messagesAdded.push(...messagesAdded);
+        if (messagesDeleted) acc.messagesDeleted.push(...messagesDeleted);
+
+        return acc;
+      },
+      { messagesAdded: [], messagesDeleted: [] },
+    );
+
+    const uniqueMessagesAdded = messagesAdded.filter(
+      (messageId) => !messagesDeleted.includes(messageId),
+    );
+
+    const uniqueMessagesDeleted = messagesDeleted.filter(
+      (messageId) => !messagesAdded.includes(messageId),
+    );
+
+    return {
+      messagesAdded: uniqueMessagesAdded,
+      messagesDeleted: uniqueMessagesDeleted,
+    };
+  }
+}
