+++ packages/twenty-server/src/modules/messaging/services/gmail-partial-message-list-fetch/gmail-partial-message-list-fetch-v2.service.ts
@@ -12,14 +12,13 @@ import {
   MessageChannelSyncStatus,
   MessageChannelSyncSubStatus,
 } from 'src/modules/messaging/standard-objects/message-channel.workspace-entity';
-import { GMAIL_USERS_HISTORY_MAX_RESULT } from 'src/modules/messaging/constants/gmail-users-history-max-result.constant';
-import { GmailError } from 'src/modules/messaging/types/gmail-error';
 import { CacheStorageService } from 'src/engine/integrations/cache-storage/cache-storage.service';
 import { InjectCacheStorage } from 'src/engine/integrations/cache-storage/decorators/cache-storage.decorator';
 import { CacheStorageNamespace } from 'src/engine/integrations/cache-storage/types/cache-storage-namespace.enum';
 import { MessageChannelMessageAssociationWorkspaceEntity } from 'src/modules/messaging/standard-objects/message-channel-message-association.workspace-entity';
 import { MessageChannelMessageAssociationRepository } from 'src/modules/messaging/repositories/message-channel-message-association.repository';
 import { GmailPartialMessageListFetchErrorHandlingService } from 'src/modules/messaging/services/gmail-partial-message-list-fetch/gmail-partial-message-list-fetch-error-handling.service';
+import { GmailGetHistoryService } from 'src/modules/messaging/services/gmail-partial-message-list-fetch/gmail-get-history.service';
 
 @Injectable()
 export class GmailPartialMessageListFetchV2Service {
@@ -40,6 +39,7 @@ export class GmailPartialMessageListFetchV2Service {
     )
     private readonly messageChannelMessageAssociationRepository: MessageChannelMessageAssociationRepository,
     private readonly gmailPartialMessageListFetchErrorHandlingService: GmailPartialMessageListFetchErrorHandlingService,
+    private readonly gmailGetHistoryService: GmailGetHistoryService,
   ) {}
 
   public async processMessageListFetch(
@@ -129,10 +129,11 @@ export class GmailPartialMessageListFetchV2Service {
     const gmailClient: gmail_v1.Gmail =
       await this.gmailClientProvider.getGmailClient(refreshToken);
 
-    const { history, historyId, error } = await this.getHistoryFromGmail(
-      gmailClient,
-      lastSyncHistoryId,
-    );
+    const { history, historyId, error } =
+      await this.gmailGetHistoryService.getHistory(
+        gmailClient,
+        lastSyncHistoryId,
+      );
 
     if (error) {
       await this.gmailPartialMessageListFetchErrorHandlingService.handleGmailError(
@@ -172,7 +173,7 @@ export class GmailPartialMessageListFetchV2Service {
     }
 
     const { messagesAdded, messagesDeleted } =
-      await this.getMessageIdsFromHistory(history);
+      await this.gmailGetHistoryService.getMessageIdsFromHistory(history);
 
     await this.cacheStorage.setAdd(
       `messages-to-import:${workspaceId}:gmail:${gmailMessageChannel.id}`,
@@ -212,98 +213,5 @@ export class GmailPartialMessageListFetchV2Service {
       MessageChannelSyncSubStatus.MESSAGES_IMPORT_PENDING,
       workspaceId,
     );
-  }
-
-  private async getMessageIdsFromHistory(
-    history: gmail_v1.Schema$History[],
-  ): Promise<{
-    messagesAdded: string[];
-    messagesDeleted: string[];
-  }> {
-    const { messagesAdded, messagesDeleted } = history.reduce(
-      (
-        acc: {
-          messagesAdded: string[];
-          messagesDeleted: string[];
-        },
-        history,
-      ) => {
-        const messagesAdded = history.messagesAdded?.map(
-          (messageAdded) => messageAdded.message?.id || '',
-        );
-
-        const messagesDeleted = history.messagesDeleted?.map(
-          (messageDeleted) => messageDeleted.message?.id || '',
-        );
-
-        if (messagesAdded) acc.messagesAdded.push(...messagesAdded);
-        if (messagesDeleted) acc.messagesDeleted.push(...messagesDeleted);
-
-        return acc;
-      },
-      { messagesAdded: [], messagesDeleted: [] },
-    );
-
-    const uniqueMessagesAdded = messagesAdded.filter(
-      (messageId) => !messagesDeleted.includes(messageId),
-    );
-
-    const uniqueMessagesDeleted = messagesDeleted.filter(
-      (messageId) => !messagesAdded.includes(messageId),
-    );
-
-    return {
-      messagesAdded: uniqueMessagesAdded,
-      messagesDeleted: uniqueMessagesDeleted,
-    };
-  }
-
-  private async getHistoryFromGmail(
-    gmailClient: gmail_v1.Gmail,
-    lastSyncHistoryId: string,
-  ): Promise<{
-    history: gmail_v1.Schema$History[];
-    historyId?: string | null;
-    error?: GmailError;
-  }> {
-    const fullHistory: gmail_v1.Schema$History[] = [];
-    let pageToken: string | undefined;
-    let hasMoreMessages = true;
-    let nextHistoryId: string | undefined;
-
-    while (hasMoreMessages) {
-      try {
-        const response = await gmailClient.users.history.list({
-          userId: 'me',
-          maxResults: GMAIL_USERS_HISTORY_MAX_RESULT,
-          pageToken,
-          startHistoryId: lastSyncHistoryId,
-          historyTypes: ['messageAdded', 'messageDeleted'],
-        });
-
-        nextHistoryId = response?.data?.historyId ?? undefined;
-
-        if (response?.data?.history) {
-          fullHistory.push(...response.data.history);
-        }
-
-        pageToken = response?.data?.nextPageToken ?? undefined;
-        hasMoreMessages = !!pageToken;
-      } catch (error) {
-        const errorData = error?.response?.data?.error;
-
-        if (errorData) {
-          return {
-            history: [],
-            error: errorData,
-            historyId: lastSyncHistoryId,
-          };
-        }
-
-        throw error;
-      }
-    }
-
-    return { history: fullHistory, historyId: nextHistoryId };
   }
 }
